# docs/swagger.yaml
definitions:
  models.Comment:
    properties:
      body:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      postId:
        description: foreign key
        type: integer
    type: object
  models.Post:
    properties:
      body:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nix Education Trainee Task API
  version: "1.0"
paths:
  /restricted/comments:
    get:
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
      summary: Get all comments
  /restricted/posts:
    get:
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: Get all posts
  /restricted/posts/{id}:
    get:
      parameters:
      - description: Post Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Get post by id
  /restricted/posts/{id}/comments/add:
    post:
      parameters:
      - description: post Id
        in: formData
        name: postid
        required: true
        type: integer
      - description: Name
        in: formData
        name: name
        required: true
        type: string
      - description: E-mail
        in: formData
        name: email
        required: true
        type: string
      - description: Body
        in: formData
        name: body
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      summary: Add comment for given post Id
  /restricted/posts/add:
    post:
      parameters:
      - description: User Id
        in: formData
        name: userid
        required: true
        type: integer
      - description: Post Title
        in: formData
        name: title
        required: true
        type: string
      - description: Post Body
        in: formData
        name: body
        required: true
        type: string
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Add post
swagger: "2.0"